Extract most output functions into the output module.
