+++ Vic2ToHoI4/Source/HOI4World/HoI4Localisation.cpp
@@ -975,129 +975,3 @@ void HoI4Localisation::insertScriptedLocalisation(
 	text += "\t\tlocalization_key = " + localisationKey + replacementKey + "\n";
 	scriptedLocalisation.addText(text);
 }
-
-
-void HoI4Localisation::Output() const
-{
-	LOG(LogLevel::Debug) << "Writing localisations";
-	const auto localisationPath = "output/" + theConfiguration.getOutputName() + "/localisation";
-	if (!Utils::TryCreateFolder(localisationPath))
-	{
-		LOG(LogLevel::Error) << "Could not create localisation folder";
-		exit(-1);
-	}
-
-	outputCountries(localisationPath);
-	outputFocuses(localisationPath);
-	outputStateLocalisations(localisationPath);
-	outputVPLocalisations(localisationPath);
-	outputIdeaLocalisations(localisationPath);
-	outputEventLocalisations(localisationPath);
-	outputPoliticalPartyLocalisations(localisationPath);
-	outputDecisionLocalisations(localisationPath);
-	outputCustomLocalisations(localisationPath);
-}
-
-
-void HoI4Localisation::outputCountries(const string& localisationPath) const
-{
-	outputLocalisations(localisationPath + "/countries_mod_l_", countryLocalisations);
-}
-
-
-void HoI4Localisation::outputFocuses(const string& localisationPath) const
-{
-	outputLocalisations(localisationPath + "/focus_mod_l_", newFocuses);
-}
-
-
-void HoI4Localisation::outputStateLocalisations(const string& localisationPath) const
-{
-	for (const auto& languageToLocalisations: stateLocalisations)
-	{
-		if (languageToLocalisations.first == "")
-		{
-			continue;
-		}
-		ofstream localisationFile(
-			localisationPath + "/state_names_l_" + languageToLocalisations.first + ".yml",
-			ios_base::app
-		);
-		if (!localisationFile.is_open())
-		{
-			LOG(LogLevel::Error) << "Could not update state localisation text file";
-			exit(-1);
-		}
-		localisationFile << "\xEF\xBB\xBF"; // output a BOM to make HoI4 happy
-		localisationFile << "l_" << languageToLocalisations.first << ":\n";
-
-		for (const auto& mapping : languageToLocalisations.second)
-		{
-			localisationFile << " STATE_" << mapping.first << ":10 \"" << mapping.second << "\"" << endl;
-		}
-	}
-}
-
-
-void HoI4Localisation::outputVPLocalisations(const string& localisationPath) const
-{
-	outputLocalisations(localisationPath + "/victory_points_l_", VPLocalisations);
-}
-
-
-void HoI4Localisation::outputIdeaLocalisations(const string& localisationPath) const
-{
-	outputLocalisations(localisationPath + "/converted_ideas_l_", ideaLocalisations);
-}
-
-
-void HoI4Localisation::outputEventLocalisations(const string& localisationPath) const
-{
-	outputLocalisations(localisationPath + "/converted_events_l_", newEventLocalisations);
-}
-
-
-void HoI4Localisation::outputPoliticalPartyLocalisations(const string& localisationPath) const
-{
-	outputLocalisations(localisationPath + "/parties3_l_", politicalPartyLocalisations);
-}
-
-
-void HoI4Localisation::outputDecisionLocalisations(const std::string& localisationPath) const
-{
-	outputLocalisations(localisationPath + "/decisions3_l_", decisionLocalisations);
-}
-
-
-void HoI4Localisation::outputCustomLocalisations(const std::string& localisationPath) const
-{
-	outputLocalisations(localisationPath + "/custom_localization_converter_l_", customLocalisations);
-}
-
-
-void HoI4Localisation::outputLocalisations(
-	const string& filenameStart,
-	const languageToLocalisationsMap& localisations
-)
-{
-	for (const auto& languageToLocalisations: localisations)
-	{
-		if (languageToLocalisations.first == "")
-		{
-			continue;
-		}
-		ofstream localisationFile(filenameStart + languageToLocalisations.first + ".yml", ios_base::app);
-		if (!localisationFile.is_open())
-		{
-			LOG(LogLevel::Error) << "Could not update localisation text file";
-			exit(-1);
-		}
-		localisationFile << "\xEF\xBB\xBF"; // output a BOM to make HoI4 happy
-		localisationFile << "l_" << languageToLocalisations.first << ":\n";
-
-		for (const auto& mapping: languageToLocalisations.second)
-		{
-			localisationFile << " " << mapping.first << ":10 \"" << mapping.second << "\"" << endl;
-		}
-	}
-}
