+++ Vic2ToHoI4/Source/HOI4World/Technologies.cpp
@@ -91,45 +91,3 @@ bool HoI4::technologies::hasTechnology(const std::string& technology) const
 {
 	return (mainTechnologies.count(technology) > 0) || (nonMtgNavalTechnologies.count(technology) > 0) || (mtgNavalTechnologies.count(technology) > 0);
 }
-
-
-void HoI4::technologies::outputTechnology(std::ostream& output) const
-{
-	output << "# Starting tech\n";
-	output << "set_technology = {\n";
-	for (auto technology: mainTechnologies)
-	{
-		output << "\t" << technology << " = 1\n";
-	}
-	output << "}\n";
-	output << "\n";
-	output << "if = {\n";
-	output << "\tlimit = { not = { has_dlc = \"Man the Guns\" } }\n";
-	output << "\tset_technology = {\n";
-	for (auto technology: nonMtgNavalTechnologies)
-	{
-		output << "\t\t" << technology << " = 1\n";
-	}
-	output << "\t}\n";
-	output << "}\n";
-	output << "if = {\n";
-	output << "\tlimit = { has_dlc = \"Man the Guns\" }\n";
-	output << "\tset_technology = {\n";
-	for (auto technology: mtgNavalTechnologies)
-	{
-		output << "\t\t" << technology << " = 1\n";
-	}
-	output << "\t}\n";
-	output << "}\n";
-}
-
-
-void HoI4::technologies::outputResearchBonuses(std::ostream& output) const
-{
-	output << "# Research Bonuses\n";
-	for (auto researchBonus : researchBonuses)
-	{
-		std::string name = researchBonus.first.substr(0, researchBonus.first.size() - 5) + "_bonus";
-		output << "add_tech_bonus = { name = " << name << " bonus = 0." << float(researchBonus.second) << " uses = 1 category = " << researchBonus.first << " }\n";
-	}
-}
