+++ Vic2ToHoI4/Source/OutHoi4/OutFocus.h
@@ -1,0 +1,14 @@
+#ifndef OUT_FOCUS_H
+#define OUT_FOCUS_H
+
+
+
+#include "../HOI4World/HoI4Focus.h"
+
+
+
+std::ostream& operator<<(std::ostream& output, const HoI4Focus& focus);
+
+
+
+#endif // OUT_FOCUS_H
