+++ Vic2ToHoI4/Source/OutHoi4/OutHoI4Version.cpp
@@ -1,0 +1,11 @@
+#include "OutHoI4Version.h"
+#include <ostream>
+
+
+
+std::ostream& HoI4::operator<<(std::ostream& out, const HoI4::Version& version)
+{
+	out << version.first << '.' << version.second << '.' << version.third;
+
+	return out;
+}
