+++ Vic2ToHoI4/Source/OutHoi4/OutHoI4Version.h
@@ -1,0 +1,19 @@
+#ifndef OUT_HOI4_VERSION_H
+#define OUT_HOI4_VERSION_H
+
+
+
+#include "../HOI4World/HOI4Version.h"
+
+
+
+namespace HoI4
+{
+
+std::ostream& operator<<(std::ostream& out, const Version& version);
+
+}
+
+
+
+#endif // OUT_HOI4_VERSION_H
