+++ Vic2ToHoI4/Source/OutHoi4/OutLocalisation.h
@@ -1,0 +1,20 @@
+#ifndef OUT_LOCALISATION_H
+#define OUT_LOCALISATION_H
+
+
+
+#include "../HOI4World/HoI4Localisation.h"
+#include "../Configuration.h"
+
+
+
+namespace HoI4
+{
+
+void outputLocalisation(const HoI4Localisation& localisation, const Configuration& theConfiguration);
+
+}
+
+
+
+#endif // OUT_LOCALISATION_H
