+++ Vic2ToHoI4/Source/OutHoi4/OutOnActions.h
@@ -1,0 +1,23 @@
+#ifndef OUT_ON_ACTIONS_H
+#define OUT_ON_ACTIONS_H
+
+
+
+#include "../Configuration.h"
+#include "../HOI4World/OnActions.h"
+#include <set>
+
+
+
+namespace HoI4
+{
+
+void outputOnActions(const OnActions& onActions,
+	 const std::set<std::string>& majorIdeologies,
+	 const Configuration& theConfiguration);
+
+}
+
+
+
+#endif // OUT_ON_ACTIONS_H
